+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/jupiter/Spend.java
@@ -10,7 +10,7 @@ import java.lang.annotation.Target;
 
 @Target(ElementType.METHOD)
 @Retention(RetentionPolicy.RUNTIME)
-@ExtendWith({SpendExtension.class, SpendExtensionParameterResolver.class})
+@ExtendWith({SpendExtension.class})
 public @interface Spend {
 
     String username();
