+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/jupiter/SpendExtension.java
@@ -2,15 +2,15 @@ package guru.qa.niffler.jupiter;
 
 import guru.qa.niffler.api.SpendService;
 import guru.qa.niffler.model.SpendJson;
+import io.qameta.allure.AllureId;
 import okhttp3.OkHttpClient;
-import org.junit.jupiter.api.extension.BeforeEachCallback;
-import org.junit.jupiter.api.extension.ExtensionContext;
+import org.junit.jupiter.api.extension.*;
 import retrofit2.Retrofit;
 import retrofit2.converter.jackson.JacksonConverterFactory;
 
 import java.util.Date;
 
-public class SpendExtension implements BeforeEachCallback {
+public class SpendExtension implements BeforeEachCallback, ParameterResolver {
 
     public static ExtensionContext.Namespace NAMESPACE = ExtensionContext.Namespace.create(SpendExtension.class);
 
@@ -35,7 +35,30 @@ public class SpendExtension implements BeforeEachCallback {
             spend.setSpendDate(new Date());
             spend.setCurrency(annotation.currency());
             SpendJson createdSpend = spendService.addSpend(spend).execute().body();
-            context.getStore(NAMESPACE).put("spend", createdSpend);
+            context.getStore(NAMESPACE).put(getAllureId(context), createdSpend);
         }
     }
+
+    @Override
+    public boolean supportsParameter(ParameterContext parameterContext, ExtensionContext extensionContext) throws ParameterResolutionException {
+        return parameterContext.getParameter()
+                .getType()
+                .isAssignableFrom(SpendJson.class);
+    }
+
+    @Override
+    public SpendJson resolveParameter(ParameterContext parameterContext, ExtensionContext extensionContext) throws ParameterResolutionException {
+        return extensionContext
+                .getStore(SpendExtension.NAMESPACE)
+                .get(getAllureId(extensionContext), SpendJson.class);
+    }
+
+    static String getAllureId(ExtensionContext context) {
+        AllureId allureId = context.getRequiredTestMethod().getAnnotation(AllureId.class);
+        if (allureId == null) {
+            throw new IllegalStateException("Annotation @AllureId must be present!");
+        }
+        return allureId.value();
+    }
+
 }
